weather_pipeline/requirements.txt
streamlit
requests
python-dotenv
schedule
pandas
pytest